Fixed the 'beans_path_to_url' and 'beans_url_to_path' functions.
beans_path_to_url() and beans_url_to_path() should be stable on all
types of server setup now. We don’t use $_SERVER variables on purpose
to make it as solid as possible. We also removed beans_abspath() which
doesn’t need BC since it was never pushed to a stable version.