Optimized beans_add_action.
As we are assigning the values to the right keys, we will not limit someone
from assigning "null" to any of the parameters. Instead, we will rely on
the developer to know how to add an action.

Shifting that design approach allows us to remove the _beans_is_action_valid
check at the end of the function.