Changed "Posts" to "Page" for prev/next text.
The word "posts" only fits when the content on the web page is a
"post" post type.  But when this content is a custom post type,
such as Listings, Books, Discussions, etc., the word "Posts"
is confusing and misleading.

The word "previous" literally means to go to the previous. In
this case, it's a web page.  Same with the word "next".

Therefore, we can standardize on the word "Page" to generally
describing going to the previous or next web page.