Refactored to instantiate through callback.
The v1.4.x design is difficult to nearly impossible to test.

Why?

1. We can't grab the instance of the object.
2. We can't reload the file for subsequent tests, i.e. as it would
cause a fatal error for reloading the same class into memory.

--------------------
How do we fix this?
--------------------

First of all, we have to return the instance of the object.  That
solves issue #1.

For issue #2, we have a couple of choices:

1. Wrap the entire class in a `class_exists()` conditional to avoid
a fatal error upon reloading the file. (This is less readable.)
2. Create the object outside of the class file.

-----------------
Proposed solution
-----------------

This commit uses option #2 above.

Instead of automatically instantiating upon file load, added
a new callback, that is hooked into after the compiler components
are loaded into memory.