Abstracted getting the asset dependencies' sources.
Abstracted the inner loop for getting the asset dependencies' sources
to a separate private method.  Why? To reduce complexity in the
compile_enqueued() method.  There's just too much going on in that
method.

Note, the collected asset sources is passed by reference. Why?
We do this to reduce memory when passing around an array.  It allows
the new method to add additional assets to the collection without
having to merge those sources back in when returning.