Made the "skip" conditional more readable.
The inline comment existed to explain the intent of the conditional,
i.e. to determine if the asset's handle should not be compiled.

This change moves that conditional to a private method.  Using the
name of the method, we can quickly understand what the check is
doing and why.

This change makes the code more readable.