Optimized to process asset dependencies once.
The previous design iterated through all assets and each asset's dependencies.
If more than one asset used the same dependency, that dependency was
processed more than once.

For example, let's say we have 5 scripts and each of the require 'jquery'
as a dependency.  jQuery requires 2 other dependencies: 'jquery-core' and
'jquery-migrate'.

The compile_enqueued() processed these dependencies 5 times, once for each
script that specified it as a dependency.  Wow, that's a lot of PHP
cycles and memory consumption.

We only need to process a dependency once.

This change keeps track of each handle that is processed.  If a handle
has already been processed, it skips over the processing code.

The results:

1. The Page Compiler is faster.
2. It uses less memory.