Made _Beans_Uikit::to_filename() more readable and optimized.
1. By moving the first guard clause to the top, we eliminate needing to create the ignore
array and variable in memory. Results:
a. That gives us a slight improvement in memory and processing time.
b. It's more readable as the intent is clear.

2. Abstracted the ignore checker to a private function. Result:
a. Makes the intent more clear when used.
b. It's more readable and understandable.
c. It's eliminates the variable assignment to give us a slight improvement in both
memory and processing time.

3. Moved the ignored components array to a private property.  This step eliminates the need
to recreate the temporary array over and over again whenever `ignore_component()` is called.