Refactored to move processing dependencies to new functions.
There might be instances when a plugin or theme developer will need to get
all of the dependencies.  The new beans_uikit_get_all_dependencies() function
gives them access to the private method.

For processing the dependencies and loading them into the registry, this commit
adds a new private function _beans_uikit_autoload_dependencies().  This
refactoring makes the enqueue function easier to read and understand, while
separating the task of dependencies out.