Optimized dequeuing all components.
When $components is true, the intent is to reset the registry by removing
all of the components.  Instead of calling beans_uikit_get_all_components(),
we simply assign an empty array to the global and then bail out.

Why? beans_uikit_get_all_components() is expensive as it iterates the filesystem,
runs several array functions that rebuilds arrays (i.e. array map, merge, unique, etc.),
and iterates through all the directories.

This commit is faster and uses less memory.