Added beans_str_ends_with().
Often the code, we need to check if a string ends with some substring, such as
ending with an extension.  The code can be confusing to understanding.

This commit adds a new utility function that accepts a string to check and
one or more substring(s) to check.

It uses mb_ for different languages.