Added beans_str_starts_with().
Like the beans_str_ends_with() function, this function makes our code more
readable without comments.  We are communicating that the conditional is
checking to find the given string pattern at the start.

An example is in the UIkit beans_uikit_register_theme().  We check to see
with the path starts with http://.  If yes, then it needs to be converted to
a path instead of a URI.

Adding this function gives developers a handy string utility.